+++ src/darker/tests/test_main.py
@@ -5,6 +5,7 @@
 import logging
 import os
 import re
+from argparse import ArgumentError
 from pathlib import Path
 from types import SimpleNamespace
 from unittest.mock import Mock, patch
@@ -14,9 +15,9 @@ from black import find_project_root
 
 import darker.__main__
 import darker.import_sorting
-from darker.git import RevisionRange
+from darker.git import WORKTREE, RevisionRange
 from darker.tests.helpers import isort_present
-from darker.utils import TextDocument
+from darker.utils import TextDocument, joinlines
 from darker.verification import NotEquivalentError
 
 
@@ -203,6 +204,88 @@ def test_format_edited_parts_ast_changed(git_repo, caplog):
     ]
 
 
+def test_format_edited_parts_isort_on_already_formatted(git_repo):
+    """An already correctly formatted file after ``isort`` is simply skipped"""
+    before = [
+        "import a",
+        "import b",
+        "",
+        "a.foo()",
+        "b.bar()",
+    ]
+    after = [
+        "import b",
+        "",
+        "b.bar()",
+    ]
+    paths = git_repo.add({"a.py": joinlines(before)}, commit="Initial commit")
+    paths["a.py"].write_text(joinlines(after))
+
+    result = darker.__main__.format_edited_parts(
+        git_repo.root,
+        {Path("a.py")},
+        RevisionRange("HEAD"),
+        enable_isort=True,
+        black_args={},
+    )
+
+    assert list(result) == []
+
+
+@pytest.mark.kwparametrize(
+    dict(rev1="HEAD^", rev2="HEAD", expect=[]),
+    dict(rev1="HEAD^", rev2=WORKTREE, expect=[(":WORKTREE:", "reformatted")]),
+    dict(rev1="HEAD", rev2=WORKTREE, expect=[(":WORKTREE:", "reformatted")]),
+)
+def test_format_edited_parts_historical(git_repo, rev1, rev2, expect):
+    """``format_edited_parts()`` is correct for different commit pairs"""
+    a_py = {
+        "HEAD^": TextDocument.from_lines(
+            [
+                "import a",
+                "from b import bar, foo",
+                "",
+                "a.foo()",
+                "bar()",
+            ]
+        ),
+        "HEAD": TextDocument.from_lines(
+            [
+                "from b import bar, foo",
+                "",
+                "bar()",
+            ]
+        ),
+        ":WORKTREE:": TextDocument.from_lines(
+            [
+                "from b import foo, bar",
+                "",
+                "bar( )",
+            ]
+        ),
+        "reformatted": TextDocument.from_lines(
+            [
+                "from b import bar, foo",
+                "",
+                "bar()",
+            ]
+        ),
+    }
+    paths = git_repo.add({"a.py": a_py["HEAD^"].string}, commit="Initial commit")
+    git_repo.add({"a.py": a_py["HEAD"].string}, commit="Modified a.py")
+    paths["a.py"].write_text(a_py[":WORKTREE:"].string)
+
+    result = darker.__main__.format_edited_parts(
+        git_repo.root,
+        {Path("a.py")},
+        RevisionRange(rev1, rev2),
+        enable_isort=True,
+        black_args={},
+    )
+
+    assert list(result) == [(paths["a.py"], a_py[x[0]], a_py[x[1]]) for x in expect]
+
+
 @pytest.mark.kwparametrize(
     dict(
         arguments=["--diff"],
@@ -293,6 +376,13 @@ def test_main_encoding(git_repo, encoding, text, newline):
     result = paths["a.py"].read_bytes()
     assert retval == 0
     assert result == b"".join(expect)
+
+
+def test_main_historical():
+    """Stop if rev2 isn't the working tree and no ``--diff`` or ``--check`` provided"""
+    with pytest.raises(ArgumentError):
+
+        darker.__main__.main(["--revision=foo..bar"])
 
 
 def test_output_diff(tmp_path, monkeypatch, capsys):