+++ CHANGES.rst
@@ -5,6 +5,7 @@ These features will be included in the next release:
 
 Added
 -----
+- Upgrade to and satisfy MyPy 0.910 by adding ``types-toml`` as a test dependency.
 
 Fixed
 -----