+++ CHANGES.rst
@@ -5,7 +5,8 @@ These features will be included in the next release:
 
 Added
 -----
-- Upgrade to and satisfy MyPy 0.910 by adding ``types-toml`` as a test dependency.
+- Upgrade to and satisfy MyPy 0.910 by adding ``types-toml`` as a test dependency, and
+  ``types-dataclasses`` as well if running on Python 3.6.
 
 Fixed
 -----